Refactor provider switch methods in dao and tests
