+++ app/dao/provider_details_dao.py
@@ -1,9 +1,12 @@
 from datetime import datetime
 
-from flask import current_app
 from sqlalchemy import asc
 
 from app.dao.dao_utils import transactional
+from app.provider_details.switch_providers import (
+    provider_is_already_primary_or_inactive,
+    update_provider_priorities
+)
 from app.models import ProviderDetails, ProviderDetailsHistory
 from app import db
 
@@ -42,55 +45,20 @@ def get_current_provider(notification_type):
 def dao_toggle_sms_provider():
     current_provider = get_current_provider('sms')
     alternate_provider = get_alternative_sms_provider(current_provider.identifier)
-
     dao_switch_sms_provider_to_provider_with_identifier(alternate_provider.identifier)
 
 
 @transactional
 def dao_switch_sms_provider_to_provider_with_identifier(identifier):
-    # Do nothing if the provider is already primary or set as inactive
     current_provider = get_current_provider('sms')
-    if current_provider.identifier == identifier:
-        current_app.logger.warning('Provider {} is already activated'.format(current_provider.display_name))
-        return current_provider
-
     new_provider = get_provider_details_by_identifier(identifier)
-    if not new_provider.active:
-        current_app.logger.warning('Cancelling switch from {} to {} as {} is inactive'.format(
-            current_provider.identifier,
-            new_provider.identifier,
-            new_provider.identifier
-        ))
+
+    if provider_is_already_primary_or_inactive(current_provider, new_provider, identifier):
         return current_provider
-
-    # Swap priority to change primary provider
-    if new_provider.priority > current_provider.priority:
-        new_provider.priority, current_provider.priority = current_provider.priority, new_provider.priority
-        _print_provider_switch_logs(current_provider, new_provider)
-        db.session.add_all([current_provider, new_provider])
-
-    # Incease other provider priority if equal
-    elif new_provider.priority == current_provider.priority:
-        current_provider.priority += 10
-        _print_provider_switch_logs(current_provider, new_provider)
-        db.session.add(current_provider)
-
-
-def _print_provider_switch_logs(current_provider, new_provider):
-    current_app.logger.warning('Switching provider from {} to {}'.format(
-        current_provider.identifier,
-        new_provider.identifier
-    ))
-
-    current_app.logger.warning('Provider {} now updated with priority of {}'.format(
-        current_provider.identifier,
-        current_provider.priority
-    ))
-
-    current_app.logger.warning('Provider {} now updated with priority of {}'.format(
-        new_provider.identifier,
-        new_provider.priority
-    ))
+    else:
+        updated_providers = update_provider_priorities(current_provider, new_provider)
+        for provider in updated_providers:
+            db.session.add(provider)
 
 
 def get_provider_details_by_notification_type(notification_type):
